third_party/blink/web_tests/wpt_internal/overscroll-gestures/overscroll-negative-scroll.html

<!DOCTYPE html>
<html >
<title>Overscroll area: negative scroll offsets</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  html, body {
    margin: 0;
  }
  #container {
    width: 200px;
    height: 200px;
    overflow: auto;
  }
  #menu {
    width: 100%;
    height: 100%;
    right: 100%; /* Overscrolls to the left of the container. */
  }
  #content {
    height: 600px;
  }
</style>
<div id="container" overscrollcontainer>
  <div id="menu"></div>
  <div id="content"></div>
</div>
<button id=button command="toggle-overscroll" commandfor="menu"></button>
<script>
test(() => {
  // The container is only vertically scrollable
  assert_equals(container.scrollWidth, container.clientWidth);
  assert_greater_than(container.scrollHeight, container.clientHeight);

  // Actual width can vary based on whether scrollbar is overlay or not.
  assert_equals(menu.clientWidth, content.clientWidth);

  // Initially the menu is offscreen to the left.
  assert_equals(menu.getBoundingClientRect().right, 0);

  // Bring the menu into view.
  menu.scrollIntoView();
  assert_equals(menu.getBoundingClientRect().left, 0);

  // The container has not scrolled.
  assert_equals(container.offsetLeft, 0);
}, "overscrollcontainer's overscroll areas can scroll to negative offsets");
</script>
</html>
